Move transform_sitk_indices_to_physical_poiints to sitk_volume
