+++ src/aind_mri_utils/sitk_volume.py
@@ -5,6 +5,7 @@ SimpleITK example code is under Apache License, see:
 https://github.com/SimpleITK/TUTORIAL/blob/main/LICENSE
 
 """
+import numpy as np
 import SimpleITK as sitk
 
 
@@ -175,3 +176,30 @@ def resample3D(
         output_direction,
     )
     return resampled_image
+
+
+def transform_sitk_indices_to_physical_points(simage, index_arr):
+    """Transforms indices indices of simage to physical points
+
+    For a SimpleITK image `simage` and a list of indices `index_arr`, transform
+    each index to a physical point.
+
+    Parameters
+    ----------
+    simage : M-d SimpleITK image
+    index_arr : numpy.ndarray (NxM)
+        matrix of indices of `simage`, where each row is an index
+
+    Returns
+    -------
+    position_arr: numpy.ndarray (NxM)
+        matrix of physical points for each index in `index_arr`
+    """
+    position_arr = np.zeros_like(index_arr, dtype="float32")
+    npt = index_arr.shape[0]
+    for ptno in range(npt):
+        ndx = tuple(map(lambda x: x.item(), index_arr[ptno, :]))
+        position_arr[ptno, :] = simage.TransformContinuousIndexToPhysicalPoint(
+            ndx
+        )
+    return position_arr
